move usage documentation into separate file
Most consumers of forks of this repo will be readers (e.g. browsing the
repo via https://github.com) rather than writers, so the README.md
should focus on describing what's in the repo rather than how to edit
it.  But we still want to keep usage instructions for the minority who
are writers.  So put them in a separate file.  This should also reduce
the chance of conflicts when merging the latest version of the template
into a fork.